+++ CHANGES.rst
@@ -8,7 +8,7 @@ Changes
 5.3 (2023-07-08)
 ----------------
 
-- Restrict access to some attributes accessible via the ``inspect`` module.
+- Forbid using some attributes providing access to restricted Python internals.
 
 
 5.2 (2021-11-19)