+++ CHANGES.rst
@@ -8,7 +8,7 @@ Changes
 6.1 (2023-07-08)
 ----------------
 
-- Restrict access to some attributes accessible via the ``inspect`` module.
+- Forbid using some attributes providing access to restricted Python internals.
 
 
 6.0 (2022-11-03)